move functions out of start
